bugfix: 创建清洗任务时修改数据集状态；无法删除已在模板/运行任务的算子
* bugfix: 创建清洗任务时修改数据集状态；无法删除已在模板/运行任务的算子

+++ backend/services/operator-market-service/src/main/java/com/datamate/operator/application/OperatorService.java
@@ -88,6 +88,9 @@ public class OperatorService {
 
     @Transactional
     public void deleteOperator(String id) {
+        if (operatorRepo.operatorInTemplateOrRunning(id)) {
+            throw BusinessException.of(OperatorErrorCode.OPERATOR_IN_INSTANCE);
+        }
         operatorRepo.deleteOperator(id);
         relationRepo.deleteByOperatorId(id);
     }
